separated sidebar component into different files in /sidebar
