+++ frontend/src/screens/layout/sidebar/desktop-menu-header.tsx
@@ -1,8 +1,8 @@
 import { Disclosure } from "@headlessui/react";
 import { Icon } from "@iconify-icon/react";
 import React from "react";
-import { Circle } from "../../components/display/circle";
-import { classNames } from "../../utils/common";
+import { Circle } from "../../../components/display/circle";
+import { classNames } from "../../../utils/common";
 import { DropDownMenuItem } from "./desktop-menu-item";
 
 interface Props {
